Fix Crash for wrong room
The crashing was caused when users entered their name into a room when the room code doesnt exist. They still got access to the page and could send requests to the server with an invalid room-code which would break it. To fix this before the users can access the room they must have receive a successful join message from the server which will only happen if the room-code exist. If the room code doesnt exist the user will just be stuck in the display room name room and cant break the server